refactor:SpecsModelsAndProductDocstring
