+++ nisystemlink/clients/spec/models/_create_specs_request.py
@@ -1,12 +1,9 @@
+from datetime import datetime
 from typing import List, Optional
 
 from nisystemlink.clients.core import ApiError
 from nisystemlink.clients.core._uplink._json_model import JsonModel
-from nisystemlink.clients.spec.models._base_specification_response import (
-    BaseSpecificationResponse,
-)
 from nisystemlink.clients.spec.models._specification import (
-    SpecificationCreation,
     SpecificationDefinition,
     SpecificationType,
 )
@@ -33,14 +30,43 @@ class CreateSpecificationsRequest(JsonModel):
     """List of specifications to be created."""
 
 
-class CreatedSpecification(BaseSpecificationResponse, SpecificationCreation):
-    """A specification successfully created on the server."""
+class BaseSpecificationResponse(JsonModel):
+    """Base Response Model for create specs response and update specs response."""
 
-    workspace: Optional[str] = None
+    id: str
+    """The global Id of the specification."""
+
+    product_id: str
+    """Id of the product to which the specification will be associated."""
+
+    spec_id: str
+    """User provided value using which the specification will be identified.
+
+    This should be unique for a product and workspace combination.
+    """
+
+    workspace: str
     """Id of the workspace to which the specification will be associated.
 
     Default workspace will be taken if the value is not given.
     """
+
+    version: int
+    """
+    Current version of the specification.
+
+    When an update is applied, the version is automatically incremented.
+    """
+
+
+class CreatedSpecification(BaseSpecificationResponse):
+    """A specification successfully created on the server."""
+
+    created_at: datetime
+    """ISO-8601 formatted timestamp indicating when the specification was created."""
+
+    created_by: str
+    """Id of the user who created the specification."""
 
 
 class CreateSpecificationsPartialSuccess(JsonModel):
